Split out elements
